Extract FileIOWriteMMAPBenchmarks into its own file.
Extracting the MMAP benchmarks into their own files allows for faster and more specific benchmark runs and makes the code leaner. This is especially relevant as we will introduce extra complexity with threaded benchmarks soon.
To not duplicate the Setup(), TearDown() and sanity_check() code it was extracted into the FileIOWriteBenchmark header. The nice way of only declaring those methods in the header and then implementing them in the corresponding cpp and including header and cpp for other build targets that need the code did not work. The benchmarks which get run by google benchmark in the
file_io_write_micro_benchmark.cpp get defined unbounded and including the cpp in any build target will automatically run them. Therefore to be able to have the implementation of SetUp() and co without having to run unwanted benchmarks, the implementation of said methods was put into the header file. This isn't pretty but as we are sure the header will only be included in our dedicated benchmark files I think it's acceptable.
