+++ src/benchmark/CMakeLists.txt
@@ -74,6 +74,7 @@ add_executable(
         micro_benchmark_main.cpp
         micro_benchmark_utils.cpp
         micro_benchmark_utils.hpp
+        file_io_write_micro_benchmark.hpp
         file_io_write_micro_benchmark.cpp
 )
 
@@ -89,6 +90,34 @@ target_link_libraries(
 # https://github.com/google/benchmark/commit/926f61da9ac8d0100eb75a5246b45484cc9c94b7
 target_link_libraries_system(
         hyriseMicroIOWriteBenchmark
+
+        benchmark::benchmark
+)
+
+
+add_executable(
+        hyriseMicroIOWriteMmapBenchmark
+
+        micro_benchmark_basic_fixture.cpp
+        micro_benchmark_basic_fixture.hpp
+        micro_benchmark_main.cpp
+        micro_benchmark_utils.cpp
+        micro_benchmark_utils.hpp
+        file_io_write_micro_benchmark.hpp
+        file_io_write_mmap_benchmark.cpp
+)
+
+target_link_libraries(
+        hyriseMicroIOWriteMmapBenchmark
+        PRIVATE
+
+        hyrise
+        hyriseBenchmarkLib
+)
+# Ignore -Wshift-sign-overflow of google benchmark introduced with
+# https://github.com/google/benchmark/commit/926f61da9ac8d0100eb75a5246b45484cc9c94b7
+target_link_libraries_system(
+        hyriseMicroIOWriteMmapBenchmark
 
         benchmark::benchmark
 )
